requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.28.0
aws-cdk.lambda-layer-kubectl-v29>=2.0.0
